+++ packages/forms/src/enhancePost.ts
@@ -4,7 +4,7 @@ import type { SubmitFunction } from '@sveltejs/kit';
 const defaultCallback: ReturnType<SubmitFunction> = async ({ result }) => applyAction(result);
 
 /** Wraps Sveltekit's `enhance` action to cancel in-flight request
- * (by using the current AbortController) if the form is re-submitted to quickly
+ * if the form is re-submitted too quickly by using the AbortController
  */
 export const enhancePost = (form: HTMLFormElement, submit?: SubmitFunction) => {
 	let previousController: AbortController;