+++ packages/forms/src/translations/newsletter-form/fr.ts
@@ -1,0 +1,9 @@
+import type { NewsletterFormTranslations } from './en.js';
+
+export const fr: NewsletterFormTranslations = {
+	emailError: 'Veuillez entrer une addresse courriel valide',
+	srLabels: {
+		submitting: 'Envoi en cours',
+		success: 'Succès!'
+	}
+};