+++ packages/forms/src/validateHoneypot.ts
@@ -1,0 +1,17 @@
+import { error, type RequestEvent } from '@sveltejs/kit';
+import { z } from 'zod';
+
+/**
+ * A function to validate the honeypot field in a form submission.
+ * For it to work, the form form must use the [Honeypot.svelte](../../components/ui/forms/Honeypot.svelte) component.
+ */
+export const validateHoneypot = async (event: RequestEvent) => {
+	const schema = z.string().max(0).optional().nullable();
+	const request = event.request.clone();
+	const data = await request.formData();
+	const honey = data.get(import.meta.env.HONEYPOT_NAME) as string;
+	const validation = schema.safeParse(honey);
+	if (!validation.success) {
+		error(404);
+	}
+};