+++ packages/components/html-elements/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @288-toolkit/html-elements
 
+## 3.0.11
+
+### Patch Changes
+
+- @288-toolkit/format@4.0.9
+
 ## 3.0.10
 
 ### Patch Changes